requirements.txt
gte-py
aiohttp
websockets
python-dotenv
async-timeout
numpy
